move triggers into separate hooks
