+++ src/components/ChatAutoComplete/ChatAutoComplete.tsx
@@ -1,19 +1,21 @@
-import React, { useCallback, useMemo, useState } from 'react';
-import debounce from 'lodash.debounce';
+import React, { useCallback } from 'react';
 
 import { AutoCompleteTextarea } from '../AutoCompleteTextarea';
-import { CommandItem, CommandItemProps } from '../CommandItem/CommandItem';
-import { EmoticonItem, EmoticonItemProps } from '../EmoticonItem/EmoticonItem';
+import type { CommandItemProps } from '../CommandItem/CommandItem';
+import type { EmoticonItemProps } from '../EmoticonItem/EmoticonItem';
 import { LoadingIndicator } from '../Loading/LoadingIndicator';
-import { UserItem, UserItemProps } from '../UserItem/UserItem';
+import type { UserItemProps } from '../UserItem/UserItem';
 
-import { useChannelStateContext } from '../../context/ChannelStateContext';
+import useEmojiIndex from './hooks/useEmojiIndex';
+import useCommandTrigger from './hooks/useCommandTrigger';
+import useEmojiTrigger from './hooks/useEmojiTrigger';
+import useUserTrigger from './hooks/useUserTrigger';
+
 import { useChatContext } from '../../context/ChatContext';
-import { useComponentContext } from '../../context/ComponentContext';
 import { useMessageInputContext } from '../../context/MessageInputContext';
 import { useTranslationContext } from '../../context/TranslationContext';
 
-import type { EmojiData, NimbleEmojiIndex } from 'emoji-mart';
+import type { EmojiData } from 'emoji-mart';
 import type {
   CommandResponse,
   UserFilters,
@@ -95,7 +97,7 @@ export type TriggerSetting<T extends UnknownType = UnknownType, U = UnknownType>
     query: string,
     text: string,
     onReady: (data: (U | undefined)[], token: string) => void,
-  ) => U[];
+  ) => U[] | Promise<void> | undefined;
   output: (
     entity: U,
   ) =>
@@ -111,12 +113,12 @@ export type TriggerSetting<T extends UnknownType = UnknownType, U = UnknownType>
 
 export type CommandTriggerSetting<
   Co extends DefaultCommandType = DefaultCommandType
-> = TriggerSetting<Partial<CommandItemProps>, SuggestionCommand<Co>>;
+> = TriggerSetting<CommandItemProps, SuggestionCommand<Co>>;
 
-export type EmojiTriggerSetting = TriggerSetting<Partial<EmoticonItemProps>, EmojiData>;
+export type EmojiTriggerSetting = TriggerSetting<EmoticonItemProps, EmojiData>;
 
 export type UserTriggerSetting<Us extends DefaultUserType<Us> = DefaultUserType> = TriggerSetting<
-  Partial<UserItemProps>,
+  UserItemProps,
   SuggestionUser<Us>
 >;
 
@@ -161,41 +163,32 @@ const UnMemoizedChatAutoComplete = <
 ) => {
   const { t } = useTranslationContext();
   const messageInput = useMessageInputContext<At, Ch, Co, Ev, Me, Re, Us, V>();
-  const commands = messageInput.getCommands();
   const {
     additionalTextareaProps,
-    onSelectItem,
-    textareaRef: innerRef,
+    autocompleteTriggers,
     disabled,
     disableMentions,
     grow,
     maxRows,
-    mentionAllAppUsers = false,
-    mentionQueryParams = {},
     SuggestionItem,
     SuggestionList,
-    autocompleteTriggers: triggers,
+    textareaRef: innerRef,
   } = messageInput;
+
+  const emojiIndex = useEmojiIndex();
+  const commandTrigger = useCommandTrigger();
+  const emojiTrigger = useEmojiTrigger(emojiIndex);
+  const userTrigger = useUserTrigger();
+
+  const triggers = autocompleteTriggers || {
+    '/': commandTrigger,
+    ':': emojiTrigger,
+    '@': userTrigger,
+  };
 
   const { onFocus, placeholder = t('Type your message'), rows = 1 } = props;
 
-  const { channel, emojiConfig } = useChannelStateContext<At, Ch, Co, Ev, Me, Re, Us>();
-  const { client, mutes } = useChatContext<At, Ch, Co, Ev, Me, Re, Us>();
-  const { EmojiIndex } = useComponentContext<At, Ch, Co, Ev, Me, Re, Us>();
-
-  const [searching, setSearching] = useState(false);
-
-  const members = channel?.state?.members;
-  const watchers = channel?.state?.watchers;
-  const { emojiData } = emojiConfig || {};
-
-  const emojiIndex: NimbleEmojiIndex | null = useMemo(() => {
-    if (EmojiIndex) {
-      // @ts-expect-error type here isn't registering the constructor type
-      return new EmojiIndex(emojiData);
-    }
-    return null;
-  }, [emojiData, EmojiIndex]);
+  const { mutes } = useChatContext<At, Ch, Co, Ev, Me, Re, Us>();
 
   const emojiReplace = (word: string) => {
     const found = emojiIndex?.search(word) || [];
@@ -206,213 +199,6 @@ const UnMemoizedChatAutoComplete = <
     if (!emoji || !('native' in emoji)) return null;
     return emoji.native;
   };
-
-  const getMembersAndWatchers = useCallback(() => {
-    const memberUsers = members ? Object.values(members).map(({ user }) => user) : [];
-    const watcherUsers = watchers ? Object.values(watchers) : [];
-    const users = [...memberUsers, ...watcherUsers];
-
-    // make sure we don't list users twice
-    const uniqueUsers = {} as Record<string, UserResponse<Us>>;
-
-    users.forEach((user) => {
-      if (user && !uniqueUsers[user.id]) {
-        uniqueUsers[user.id] = user;
-      }
-    });
-
-    return Object.values(uniqueUsers);
-  }, [members, watchers]);
-
-  const queryMembersDebounced = useCallback(
-    debounce(async (query: string, onReady: (users: UserResponse<Us>[]) => void) => {
-      try {
-        // @ts-expect-error
-        const response = await channel.queryMembers({
-          name: { $autocomplete: query },
-        });
-
-        const users = response.members.map((member) => member.user) as UserResponse<Us>[];
-
-        if (onReady && users.length) {
-          onReady(users);
-        } else {
-          onReady([]);
-        }
-      } catch (error) {
-        console.log({ error });
-      }
-    }, 200),
-    [channel],
-  );
-
-  const queryUsers = async (query: string, onReady: (users: UserResponse<Us>[]) => void) => {
-    if (!query || searching) return;
-    setSearching(true);
-
-    try {
-      const { users } = await client.queryUsers(
-        // @ts-expect-error
-        {
-          $or: [{ id: { $autocomplete: query } }, { name: { $autocomplete: query } }],
-          id: { $ne: client.userID },
-          ...mentionQueryParams.filters,
-        },
-        { id: 1, ...mentionQueryParams.sort },
-        { limit: 10, ...mentionQueryParams.options },
-      );
-
-      if (onReady && users.length) {
-        onReady(users);
-      } else {
-        onReady([]);
-      }
-    } catch (error) {
-      console.log({ error });
-    }
-
-    setSearching(false);
-  };
-
-  const queryUsersDebounced = debounce(queryUsers, 200);
-
-  /**
-   * dataProvider accepts `onReady` function, which will executed once the data is ready.
-   * Another approach would have been to simply return the data from dataProvider and let the
-   * component await for it and then execute the required logic. We are going for callback instead
-   * of async-await since we have debounce function in dataProvider. Which will delay the execution
-   * of api call on trailing end of debounce (lets call it a1) but will return with result of
-   * previous call without waiting for a1. So in this case, we want to execute onReady, when trailing
-   * end of debounce executes.
-   */
-  const getTriggers = useCallback(
-    () =>
-      triggers ||
-      ({
-        '/': {
-          component: CommandItem,
-          dataProvider: (query, text, onReady) => {
-            if (text.indexOf('/') !== 0 || !commands) {
-              return [];
-            }
-            const selectedCommands = commands.filter(
-              (command) => command.name?.indexOf(query) !== -1,
-            );
-
-            // sort alphabetically unless the you're matching the first char
-            selectedCommands.sort((a, b) => {
-              let nameA = a.name?.toLowerCase();
-              let nameB = b.name?.toLowerCase();
-              if (nameA?.indexOf(query) === 0) {
-                nameA = `0${nameA}`;
-              }
-              if (nameB?.indexOf(query) === 0) {
-                nameB = `0${nameB}`;
-              }
-              // Should confirm possible null / undefined when TS is fully implemented
-              if (nameA != null && nameB != null) {
-                if (nameA < nameB) {
-                  return -1;
-                }
-                if (nameA > nameB) {
-                  return 1;
-                }
-              }
-
-              return 0;
-            });
-
-            const result = selectedCommands.slice(0, 10);
-            if (onReady) onReady(result, query);
-
-            return result;
-          },
-          output: (entity) => ({
-            caretPosition: 'next',
-            key: entity.name,
-            text: `/${entity.name}`,
-          }),
-        },
-        ':': {
-          component: EmoticonItem,
-          dataProvider: (query, _, onReady) => {
-            if (query.length === 0 || query.charAt(0).match(/[^a-zA-Z0-9+-]/)) {
-              return [];
-            }
-            const emojis = emojiIndex?.search(query) || [];
-            // emojiIndex.search sometimes returns undefined values, so filter those out first
-            const result = emojis.filter(Boolean).slice(0, 10);
-            if (onReady) onReady(result, query);
-
-            return result;
-          },
-          output: (entity) => ({
-            caretPosition: 'next',
-            key: entity.id,
-            text: `${'native' in entity ? entity.native : ''}`,
-          }),
-        },
-        '@': {
-          callback: (item) => onSelectItem && onSelectItem(item),
-          component: UserItem,
-          dataProvider: (query, _, onReady) => {
-            if (mentionAllAppUsers) {
-              return queryUsersDebounced(query, (data: (UserResponse<Us> | undefined)[]) => {
-                if (onReady) onReady(data, query);
-              });
-            }
-
-            /**
-             * By default, we return maximum 100 members via queryChannels api call.
-             * Thus it is safe to assume, that if number of members in channel.state is < 100,
-             * then all the members are already available on client side and we don't need to
-             * make any api call to queryMembers endpoint.
-             */
-            if (!query || Object.values(members || {}).length < 100) {
-              const users = getMembersAndWatchers();
-
-              const matchingUsers = users.filter((user) => {
-                if (user.id === client.userID) return false;
-                if (!query) return true;
-
-                if (
-                  user.name !== undefined &&
-                  user.name.toLowerCase().includes(query.toLowerCase())
-                ) {
-                  return true;
-                }
-
-                return user.id.toLowerCase().includes(query.toLowerCase());
-              });
-
-              const data = matchingUsers.slice(0, 10);
-
-              if (onReady) onReady(data, query);
-
-              return data;
-            }
-
-            return queryMembersDebounced(query, (data: (UserResponse<Us> | undefined)[]) => {
-              if (onReady) onReady(data, query);
-            });
-          },
-          output: (entity) => ({
-            caretPosition: 'next',
-            key: entity.id,
-            text: `@${entity.name || entity.id}`,
-          }),
-        },
-      } as TriggerSettings<Co, Us, V>),
-    [
-      commands,
-      getMembersAndWatchers,
-      members,
-      onSelectItem,
-      queryMembersDebounced,
-      triggers,
-      emojiIndex,
-    ],
-  );
 
   const updateInnerRef = useCallback(
     (ref) => {
@@ -448,7 +234,7 @@ const UnMemoizedChatAutoComplete = <
       rows={rows}
       SuggestionItem={SuggestionItem}
       SuggestionList={SuggestionList}
-      trigger={getTriggers()}
+      trigger={triggers}
       value={messageInput.text}
     />
   );
