+++ src/components/ChatAutoComplete/hooks/useCommandTrigger.ts
@@ -1,0 +1,72 @@
+import { useMessageInputContext } from '../../../context/MessageInputContext';
+import { CommandItem } from '../../CommandItem/CommandItem';
+
+import type { CommandTriggerSetting } from '../ChatAutoComplete';
+import type {
+  DefaultAttachmentType,
+  DefaultChannelType,
+  DefaultCommandType,
+  DefaultEventType,
+  DefaultMessageType,
+  DefaultReactionType,
+  DefaultUserType,
+} from '../../../../types/types';
+
+const useCommandTrigger = <
+  At extends DefaultAttachmentType = DefaultAttachmentType,
+  Ch extends DefaultChannelType = DefaultChannelType,
+  Co extends DefaultCommandType = DefaultCommandType,
+  Ev extends DefaultEventType = DefaultEventType,
+  Me extends DefaultMessageType = DefaultMessageType,
+  Re extends DefaultReactionType = DefaultReactionType,
+  Us extends DefaultUserType<Us> = DefaultUserType
+>(): CommandTriggerSetting<Co> => {
+  const { getCommands } = useMessageInputContext<At, Ch, Co, Ev, Me, Re, Us>();
+
+  const commands = getCommands();
+
+  return {
+    component: CommandItem,
+    dataProvider: (query, text, onReady) => {
+      if (text.indexOf('/') !== 0 || !commands) {
+        return [];
+      }
+      const selectedCommands = commands.filter((command) => command.name?.indexOf(query) !== -1);
+
+      // sort alphabetically unless the you're matching the first char
+      selectedCommands.sort((a, b) => {
+        let nameA = a.name?.toLowerCase();
+        let nameB = b.name?.toLowerCase();
+        if (nameA?.indexOf(query) === 0) {
+          nameA = `0${nameA}`;
+        }
+        if (nameB?.indexOf(query) === 0) {
+          nameB = `0${nameB}`;
+        }
+        // Should confirm possible null / undefined when TS is fully implemented
+        if (nameA != null && nameB != null) {
+          if (nameA < nameB) {
+            return -1;
+          }
+          if (nameA > nameB) {
+            return 1;
+          }
+        }
+
+        return 0;
+      });
+
+      const result = selectedCommands.slice(0, 10);
+      if (onReady) onReady(result, query);
+
+      return result;
+    },
+    output: (entity) => ({
+      caretPosition: 'next',
+      key: entity.name,
+      text: `/${entity.name}`,
+    }),
+  };
+};
+
+export default useCommandTrigger;
